Inclusão de um Injetor de Dependências; finalizadas as estruturas de interfaces de query e repositórios do sistema; implementação do repositório de Comandas; inserção de DateTimeExtensions

+++ axxis.bacco.backend/axxis.bacco.backend.infra.data/Contexts/BaccoContext.cs
@@ -27,7 +27,7 @@ namespace axxis.bacco.backend.infra.data.Contexts
         public DbSet<ItemVenda> ItensVenda { get; set; }
         public DbSet<FormaPagamento> FormasPagamento { get; set; }
 
-        public double NextVal(string sequenceName)
+        public long NextVal(string sequenceName)
         {
             using (var cmd = this.Database.GetDbConnection().CreateCommand())
             {
